DocumentComponent-stuff should not depend on dialogmote
